refactor : dto 이름 및 경로 변경
Dto 이름 및 경로 변경 `util/dto/ThumbnailFileIdDto` -> `user/dto/posting/PostingImageUploadDto`

util이나 thumbnail보다는 posting에만 사용되기에 이름 및 경로를 변경했습니다.

issue : #371
pr : #401
     #discussion_r964620404
