+++ src/main/java/keeper/project/homepage/util/service/ThumbnailService.java
@@ -8,7 +8,7 @@ import java.util.stream.Stream;
 import keeper.project.homepage.exception.file.CustomInvalidImageFileException;
 import keeper.project.homepage.util.image.ImageFormatChecking;
 import keeper.project.homepage.util.MultipartFileWrapper;
-import keeper.project.homepage.util.dto.ThumbnailFileIdDto;
+import keeper.project.homepage.user.dto.posting.PostingImageUploadResponseDto;
 import keeper.project.homepage.entity.FileEntity;
 import keeper.project.homepage.entity.ThumbnailEntity;
 import keeper.project.homepage.entity.posting.PostingEntity;
@@ -189,8 +189,8 @@ public class ThumbnailService {
             .build());
   }
 
-  public ThumbnailFileIdDto getThumbnailFileId(ThumbnailEntity thumbnailEntity){
-    return ThumbnailFileIdDto.from(thumbnailEntity);
+  public PostingImageUploadResponseDto getThumbnailFileId(ThumbnailEntity thumbnailEntity){
+    return PostingImageUploadResponseDto.from(thumbnailEntity);
   }
 
   private FilePair saveFilesInServer(ThumbType type, ImagePreprocessing imagePreprocessing,
